pyproject.toml
[project]
name = "travel-planner-ai"
version = "0.1.0"
description = "Multi-agent AI travel planner with Streamlit, Agno, and GPT-4o-mini."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agno",
    "openai",
    "streamlit",
    "pydantic",
    "python-dotenv",
    "ddgs",
    "beautifulsoup4",
    "requests",
    "matplotlib",
]

[project.optional-dependencies]
dev = [
    "pytest",
]

[project.scripts]
eventaide-app = "app:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["travel_planner*"]
